Room Layouts
Rooms now spawn content depending on room type when the player enters the room for the first time. Room layouts can be created as Scriptable Objects containing a list of "TileEntry" objects which hold a position, tile, and debug tile. The debug tile goes on a separate debug tilemap for use in development to show that unimplemented objects/enemies are being spawned at the correct location. RoomLayouts are stored in Scriptable Object containers for each room type, and contain a method for getting a random room layout. More room layouts can easily be created and implemented by simply making a RoomLayout SO and placing it in the appropriate container.

Also added another step to the end of room graph generation which assigns room types. The start room is always created first and is assigned its type when created. The last room generated is always the boss room. Every other room rolls for a chance to be a treasure room, and if not a treasure room, is an enemy room.

There's now enough framework to start making room prefabs for all door positions and start generating real levels