Updated Scene
Removed unnecessary starting room GameObject, as the starting room is spawned by the LevelGenerator for neighbor connectivity.

Player GameObject starts inactive in hierarchy and player movement starts locked.

When the LevelGenerator finishes room instantiation the player is set active in the hierarchy and the starting room's shroud is lifted, and player movement is enabled once the shroud finishes fading.