Prepared Scene for Combat Testing
Disabled Level Generator and made a new GameObject "Combat Testing Room" (CTR) based from the NESW prefab.

Replaced the Room script on the CTR with a new TestingRoom script which inherits from Room and makes two simple changes:
- Calls PlayerEnteredRoom() in Awake()
- Exposes a field for a room layout which can optionally be filled to set the room's layout manually or left blank to still pick one randomly depending on room type.

Put a new very simple script on the Player which holds a list of Transforms for the room's entry point, and randomly places the player at one of those positions in Awake() to better simulate the player having just entered the room.