Finished Attacks
Updated melee attack swing to rotate overlap box as it swings and added an animation of a swinging fishing rod. Visuals has its own script to rotate to the attack direction when swung and move the animation with the player.

Implemented ranged attack, which shoots a fish projectile in the player's aimed direction with specified damage, knockback, and speed. Projectiles are destroyed upon their first collision with a non-player/player projectile collider.

Implemented a crosshair which shows up while playing on controller and aiming. The crosshair rotates around the player at a specified radius to match the direction the player aims the right stick. On MKB the player uses their mouse to aim.