Skier, Heart Drops, and Updated Art/Animations
Implemented the Skier enemy (minus sprites). It picks a random diagonal to start moving on when spawned, and then bounces around the room. They are meant to be nuisances, spawning in high numbers with high speed, but doing low damage, knockback, and hit stun (the player can now be hit stunned).

Changed the Layers and Physics2D matrix to facilitate correct collisions between the environment, player, enemies, and drops.

Enemies now have a chance to drop a heart, and a chance for that heart to be a full heart or a half heart. The chance for both is set independently for each enemy.

Updated Wall tileset, which also forced me to replace all the environment tilesets on the room prefabs. At the same time I did that, I removed the debug tilemap and added a disabled tilemap renderer set to the Details layer to each content tilemap. By assigning a square sprite and a color to the layout object tiles, toggling the content tilemap's sprite renderer has the same effect as the debug tilemap.

Added a death animation for the player and made some small changes to the player sprites to add an outline to the hood where it was missing (everything except idle and walk down) and an outline above the feet to all player sprites.

Changed the outlines of the fish projectile and rock to be soft like the player's. Hard outlines will be reserved for interactable/destructible/pickupable objects. Still need to update the fishing rod swing outlines.